feat: Search for user's personal sgoinfre directory in an optimized way: 1. Search for 1 second in common locations. 2. Search for 60 seconds from the root directory in a breadth-first and in a depth-first way in two parallel processes. 3. Prompt user to input the path manually. Once found, save the path in a `SGOINFRE` shell environment variable into all shell config files.
The `wait_for_jobs` function gets fully refactored to use it in a more modular way.

Resolves issue #22.
Mostly resolves issue #31 (except for looking for sgoinfre at installation and option to change the saved path).

Thank you to @JonatV for sparking the idea for this process in this comment: https://github.com/itislu/42free/issues/22#issuecomment-2100505422
Thank you to @loyc12 for testing the results and performance of many different approaches!