style: Make tight connection between two blocks of code more clear by removing a separator line between them
The success of `symlink` does not have to be checked in this case because it is ensured by the `mv` command directly before it.
And even if there would already be a file where the symlink should be created, it would almost certainly have been created by 42free some time before and would therefore also be a symbolic link to the target location anyway.