style: Split up line which runs an update
Bash reads scripts line-by-line, so in global scope any line that modifies the running script - like an update does - should be fully contained in one line.
However, if that line is within a function, that function has already been read into memory, so even modifications to the running script do not affect the function anymore.