improve: Make prompts more consistent and require enter key for all prompts
I noticed people tend to assume that enter is required, so they give an answer and press enter in quick succession, but this can lead to unintentionally confirming 2 prompts in a row.

Default answers when just pressing enter still differ, depending on what is the suggested or more likely answer.